requirement.txt
#fastapi==0.104.1
fastapi
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
websockets==11.0.3
python-dotenv==1.0.0
e2b_desktop==1.7.2
boto3
langchain-aws
pydantic
gradio>=4.0.0
#aiohttp>=3.9.0
aiohttp
mcp-streamablehttp-client
e2b=1.4.0